+++ firestore-send-email/functions/jest.config.js
@@ -15,7 +15,7 @@ module.exports = {
     printBasicPrototype: true,
   },
   setupFiles: ["<rootDir>/__tests__/jest.setup.ts"],
-  testMatch: ["**/__tests__/*.test.ts"],
+  testMatch: ["**/__tests__/**/*.test.ts"],
   testEnvironment: "node",
   moduleNameMapper: {
     "firebase-admin/app": "<rootDir>/node_modules/firebase-admin/lib/app",